requirements.txt
flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
transformers==4.48.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.0+cpu  # CPU-only version of PyTorch
scikit-learn==1.6.1
pdf2image==1.17.0
pytesseract==0.3.13
python-docx==1.1.2
numpy==1.24.4
requests==2.32.3
openai==1.59.7